requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper==20231117
torch==2.1.0
transformers==4.35.2
TTS==0.20.6
accelerate==0.24.1
bitsandbytes==0.41.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
python-multipart==0.0.6